Android splitting in classes, Fix audio permission, Fix crash after allow permissions
