+++ packages/camera/android/src/main/java/io/flutter/plugins/camera/CameraException.java
@@ -1,0 +1,8 @@
+package io.flutter.plugins.camera;
+
+public class CameraException extends Exception {
+
+    public CameraException(String msg) {
+        super(msg);
+    }
+}
