+++ packages/camera/android/src/main/java/io/flutter/plugins/camera/CompareSizesByArea.java
@@ -1,0 +1,15 @@
+package io.flutter.plugins.camera;
+
+import android.util.Size;
+
+import java.util.Comparator;
+
+public class CompareSizesByArea implements Comparator<Size> {
+
+    @Override
+    public int compare(Size lhs, Size rhs) {
+        // We cast here to ensure the multiplications won't overflow.
+        return Long.signum(
+                (long) lhs.getWidth() * lhs.getHeight() - (long) rhs.getWidth() * rhs.getHeight());
+    }
+}
